Cleanup: fix ~36 CodeQL alerts (warnings/errors + safe notes)
Targets actionable code-scanning alerts without touching the codebase's
intentional patterns:

- cs/unused-collection + cs/useless-assignment-to-local: remove dead local
  variables/collections (AnalyzeImpact, TraceDependency, ChatTurn).
- cs/reference-equality-with-object: ==/!= on references -> ReferenceEquals.
- cs/misleading-indentation: while(...) { if … ; … } split across lines.
- cs/useless-upcast + cs/useless-cast-to-self: remove redundant casts.
- cs/simplifiable-boolean-expression: `x is null ? true : e` -> `x is null || e`.
- cs/missed-readonly-modifier: readonly on never-reassigned collection fields.
- cs/local-not-disposed + cs/missed-using-statement: HttpRequestMessage/
  HttpContent/HttpResponseMessage wrapped in using.

Release build clean, 969/969 tests pass.

+++ Inferpal.Tests/HardwareProfileTests.cs
@@ -18,7 +18,7 @@ public class HardwareProfileTests
     public void LoadedVramBytes_SumsRunningModels()
     {
         var p = new HardwareProfile(24, [Running("a", 4), Running("b", 2)], []);
-        Assert.Equal((long)(6 * Gb), p.LoadedVramBytes);
+        Assert.Equal(6 * Gb, p.LoadedVramBytes);
         Assert.Equal(6, p.LoadedGb, precision: 1);
     }
